renaming ComponentFilesUtils.js to ComponentFilesFactory and extracting shared constants
