+++ src/config/CLI.js
@@ -1,7 +1,4 @@
 /** @type {string} **/
-export const BUILD_FOLDER_NAME = 'build'
-
-/** @type {string} **/
 export const DEV_FOLDER_NAME = '.explorer'
 
 /** @type {string} **/
