+++ src/utils/SetupFilesUtils.js
@@ -2,9 +2,13 @@
 import { basename, join, sep } from 'node:path'
 
 // Internal Dependencies
-import { SETUP_FOLDER_NAME, TEMPLATES_FOLDER_NAME, THEME_INDEX_TEMPLATE_LIQUID_FILE } from '../config/Components.js'
+import {
+  JSON_EXTENSION,
+  SETUP_FOLDER_NAME,
+  TEMPLATES_FOLDER_NAME,
+  THEME_INDEX_TEMPLATE_LIQUID_FILE
+} from '../config/Components.js'
 import { handleWatcherEvent } from './Watcher.js'
-import { JSON_EXTENSION } from '../factory/componentFilesFactory.js'
 import { copyFileAndCreatePath, getFileContents } from './FileUtils.js'
 
 const setupFolderCue = join(sep, SETUP_FOLDER_NAME, sep)
