+++ src/utils/copyrightUtils.js
@@ -1,6 +1,6 @@
 import { getPackageName } from './NodeUtils.js'
 import InternalError from '../errors/InternalError.js'
-import { FileTypes } from '../factory/componentFilesFactory.js'
+import { FileTypes } from '../config/Components.js'
 
 /**
  * Get Component File Copyright
